refactor: make arrow handler module private and drop empty pandas module
